Update Main.py
Included the feature of downloading songs from a playlist to a folder with the same name of the playlist. Edited the logger to only log errors, which are primarily an attempt to download from a unavailable URL. Fix the mistake of not including the URL when a download fails.  Modified unavailable count to include more errors.